+++ front/app/components/admin/PostManager/components/PostTable/Row/InitiativeRow.tsx
@@ -46,7 +46,7 @@ import GetAppConfiguration, {
 import GetInitiativeAllowedTransitions, {
   GetInitiativeAllowedTransitionsChildProps,
 } from 'resources/GetInitiativeAllowedTransitions';
-import { getDaysRemainingUntil } from 'utils/dateUtils';
+import { getPeriodRemainingUntil } from 'utils/dateUtils';
 
 // events
 import eventEmitter from 'utils/eventEmitter';
@@ -147,7 +147,7 @@ class InitiativeRow extends React.PureComponent<Props & WrappedComponentProps> {
       tenant.attributes.settings.initiatives
     ) {
       if (selectedStatusObject.attributes.code === 'proposed') {
-        return getDaysRemainingUntil(initiative.attributes.expires_at);
+        return getPeriodRemainingUntil(initiative.attributes.expires_at);
       } else {
         return (
           <StatusLabel